regrouped all mixture methods in a unique class and moved it into api for later use
