+++ src/main/java/tekcays_addon/gtapi/consts/ToxicFluidMaterials.java
@@ -1,12 +1,12 @@
 package tekcays_addon.gtapi.consts;
 
-import gregtech.api.unification.material.Material;
-import tekcays_addon.gtapi.unification.material.info.TKCYAMaterialFlags;
+import static gregtech.api.unification.material.Materials.*;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import static gregtech.api.unification.material.Materials.*;
+import gregtech.api.unification.material.Material;
+import tekcays_addon.gtapi.unification.material.info.TKCYAMaterialFlags;
 
 public class ToxicFluidMaterials {
