+++ src/main/java/tekcays_addon/loaders/recipe/TKCYARecipeLoader.java
@@ -96,7 +96,6 @@ public class TKCYARecipeLoader {
 
         HeatHandler.init();
         // MUST BE CALLED AFTER ANY HANDLER THAT GENERATES DUST_MIXTURE !
-        SpiralSeparatorHandler.init();
         GTCEuRequireCleanRoomHandler.init();
         // Must be called in the end
         FiltrationRecipeHandler.init();
