+++ src/main/java/tekcays_addon/loaders/recipe/chains/BauxiteChain.java
@@ -4,10 +4,10 @@ import static gregtech.api.recipes.RecipeMaps.CHEMICAL_BATH_RECIPES;
 import static gregtech.api.recipes.RecipeMaps.CHEMICAL_RECIPES;
 import static gregtech.api.unification.material.Materials.*;
 import static gregtech.api.unification.ore.OrePrefix.*;
+import static tekcays_addon.api.mixtures.MixtureRoastingRecipeHandlerMethods.getDustMixtureStackWithNBT;
 import static tekcays_addon.gtapi.consts.TKCYAValues.SECOND;
 import static tekcays_addon.gtapi.recipes.TKCYARecipeMaps.*;
 import static tekcays_addon.gtapi.unification.TKCYAMaterials.*;
-import static tekcays_addon.gtapi.utils.roasting.RoastingRecipeHandlerMethods.getDustMixtureStackWithNBT;
 import static tekcays_addon.loaders.DamageableItemsLoader.electrodeCarbon;
 
 import java.util.ArrayList;
